v0fix: fixed production log paths
Log paths are now resolved relative to the directory. Not doing so means the
absolute path gets baked in and if the path is slightly different in the actual
prod server (as compared to the build server) we get issues.

If you pass: ~~/some/log.log it will now resolve to some/log.log both during dev
and production. The logger will resolve them before passing them to pino. If you
need the full path just do path.resolve.